requirements.txt
asgiref==3.5.2
bidict==0.22.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Flask==2.1.3
Flask-Cors==3.0.10
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.1
pycparser==2.21
PySoundFile==0.9.0.post1
python-engineio==4.3.3
python-socketio==5.7.1
requests==2.28.1
scipy==1.8.1
six==1.16.0
srt==3.5.2
tqdm==4.64.0
urllib3==1.26.10
vosk==0.3.42
websockets==10.3
Werkzeug==2.1.2
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
